+++ provider.go
@@ -41,6 +41,7 @@ func Provider() *schema.Provider {
 			"proxmox_virtual_environment_version":    dataSourceVirtualEnvironmentVersion(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
+			"proxmox_virtual_environment_file":  resourceVirtualEnvironmentFile(),
 			"proxmox_virtual_environment_group": resourceVirtualEnvironmentGroup(),
 			"proxmox_virtual_environment_pool":  resourceVirtualEnvironmentPool(),
 			"proxmox_virtual_environment_role":  resourceVirtualEnvironmentRole(),