+++ CHANGELOG.md
@@ -13,6 +13,7 @@ FEATURES:
 * **New Data Source:** `proxmox_virtual_environment_user`
 * **New Data Source:** `proxmox_virtual_environment_users`
 * **New Data Source:** `proxmox_virtual_environment_version`
+* **New Resource:** `proxmox_virtual_environment_file`
 * **New Resource:** `proxmox_virtual_environment_group`
 * **New Resource:** `proxmox_virtual_environment_pool`
 * **New Resource:** `proxmox_virtual_environment_role`